+++ src/volumes/OssVolume.php
@@ -449,16 +449,11 @@ class OssVolume extends Volume
 
     /**
      * @param string $path
-     * @param bool $isRemotePath
      * @return string
      * @throws VolumeException
      */
-    protected function getRemoteObjectUrl(string $path, bool $isRemotePath = false): string
+    protected function getRemoteObjectUrl(string $path): string
     {
-        if (!$isRemotePath) {
-            $path = $this->getRemoteObjectPath($path);
-        }
-
         if ($this->hasUrls) {
             return $this->_completeSchema($this->getRootUrl()) . $this->_encodeUriPath($path);
         }